+++ state.go
@@ -732,9 +732,8 @@ func (m *Memberlist) verifyProtocol(remote []pushNodeState) error {
 			continue
 		}
 
-		// Skip nodes that don't have versions set, it just means
-		// their version is zero.
-		if len(rn.Vsn) == 0 {
+		// Skip nodes that don't have proper version info
+		if len(rn.Vsn) < 5 {
 			continue
 		}
 
@@ -783,7 +782,7 @@ func (m *Memberlist) verifyProtocol(remote []pushNodeState) error {
 	// node in the cluster satisifies this.
 	for _, n := range remote {
 		var nPCur, nDCur uint8
-		if len(n.Vsn) > 0 {
+		if len(n.Vsn) >= 6 {
 			nPCur = n.Vsn[2]
 			nDCur = n.Vsn[5]
 		}
@@ -1161,7 +1160,7 @@ func (m *Memberlist) aliveNodeLocked(a *alive, notify chan struct{}, bootstrap b
 		m.encodeBroadcastNotify(a.Node, aliveMsg, a, notify)
 
 		// Update protocol versions if it arrived
-		if len(a.Vsn) > 0 {
+		if len(a.Vsn) >= 6 {
 			state.PMin = a.Vsn[0]
 			state.PMax = a.Vsn[1]
 			state.PCur = a.Vsn[2]